requirements.txt
numpy==2.2.5
pandas==2.2.3
opencv-python==4.11.0
ultralytics==8.3.155
deep_sort_realtime==1.3.2
easyocr==1.7.2
openai==1.86.0
gTTS==2.5.4
moviepy==1.0.3
streamlit==1.45.1
elevenlabs==2.8.1
pydub==0.25.1
ffmpeg-python==0.2.0
